+++ R/html_document.R
@@ -89,6 +89,30 @@
 #'  MathJax CDN. The "local" option uses a local version of MathJax (which is
 #'  copied into the output directory). You can pass an alternate URL or pass
 #'  \code{NULL} to exclude MathJax entirely.
+#'@param math_method Math rendering engine to use. This will define the math method to use with Pandoc.
+#'
+#'  * It can be a string for the engine, one of
+#'  `r knitr::combine_words(pandoc_math_engines(), and = "or ", before = "'")`,
+#'  or `default` for `mathjax`.
+#'  * It can be a list of
+#'    * `engine`:  one of
+#'      `r knitr::combine_words(pandoc_math_engines(), and = "or ", before = "'")`.
+#'    * `url`: A specific url to use with `mathjax`, `katex` or `webtex`.
+#'      Note that for `engine = "mathjax"`, `url = "local"` will use a local version of MathJax (which is
+#'  copied into the output directory).
+#'
+#'  For example,
+#'  ```yaml
+#'  output:
+#'    html_document:
+#'      math_method:
+#'        engine: katex
+#'        url: https://cdn.jsdelivr.net/npm/katex@0.11.1/dist
+#'  ```
+#'
+#'  See [Pandoc's Manual about Math in HTML](https://pandoc.org/MANUAL.html#math-rendering-in-html) for the details.
+#'  Note that `c(mathjax = "local")` is equivalent to specifying "local" to
+#'  `mathjax`.
 #'@param section_divs Wrap sections in \code{<div>} tags, and attach identifiers to the
 #'  enclosing \code{<div>} rather than the header itself.
 #'@param template Pandoc template to use for rendering. Pass "default" to use
@@ -333,6 +357,7 @@ html_document <- function(toc = FALSE,
                           theme = "default",
                           highlight = "default",
                           highlight_downlit = FALSE,
+                          math_method = "default",
                           mathjax = "default",
                           template = "default",
                           extra_dependencies = NULL,
@@ -612,7 +637,9 @@ html_document <- function(toc = FALSE,
     on_exit = on_exit,
     base_format = html_document_base(theme = theme,
                                      self_contained = self_contained,
-                                     lib_dir = lib_dir, mathjax = mathjax,
+                                     lib_dir = lib_dir,
+                                     math_method = math_method,
+                                     mathjax = mathjax,
                                      template = template,
                                      pandoc_args = pandoc_args,
                                      extra_dependencies = extra_dependencies,
@@ -672,15 +699,6 @@ themes <- function() {
 html_highlighters <- function() {
   c(highlighters(), "textmate")
 }
-
-default_mathjax <- function() {
-  paste0("https://mathjax.rstudio.com/latest/", mathjax_config())
-}
-
-mathjax_config <- function() {
-  "MathJax.js?config=TeX-AMS-MML_HTMLorMML"
-}
-
 
 navbar_html_from_yaml <- function(navbar_yaml) {